Add files via upload
These are the implementation of code in SQL Server for apriori algorithm for finding most frequent itemsets, (like 1-itemset, 2-itemsets, etc...) 
Ex: If a person has a blood pressure, diabetes etc....,  the next frequent symptoms that a person might develop are Heart Disease or may be some combination of health issues is predicted by apriori algorithm. Binarized the data, to be used in WEKA for performing ML Algorithms to predict the hospital readmissions, most likely disability that a patient might develop, high utilize of medical services. etc..... can be answered by the approach.